requirements.txt
flask
flask-cors
google-generativeai
pillow
faiss-cpu
sentence-transformers
